+++ source/GameInterface/Services/ItemRosters/Messages/ItemRosterUpdate.cs
@@ -1,7 +1,7 @@
 ﻿using Common.Logging.Attributes;
 using Common.Messaging;
 
-namespace GameInterface.Services.ItemRosters.Messages.Events
+namespace GameInterface.Services.ItemRosters.Messages
 {
     /// <summary>
     /// Called when an item roster is updated.
@@ -11,7 +11,7 @@ namespace GameInterface.Services.ItemRosters.Messages.Events
     {
         public string PartyBaseID { get; }
         public string ItemID { get; }
-        public string ItemModifierID { get;  }
+        public string ItemModifierID { get; }
         public int Amount { get; }
 
         public ItemRosterUpdate(string partyBaseID, string itemID, string itemModifierID, int amount)
